refactor. mypage
